회원 가입 관련 로직 구현 , 로그아웃 로직 수정, Cors 설정

See merge request s09-fintech-finance-sub2/S09P22D110!51

+++ user/src/main/java/com/jrjr/inbest/login/service/LoginServiceImpl.java
@@ -5,11 +5,11 @@ import java.util.Optional;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.jrjr.inbest.global.exception.AuthenticationFailedException;
+import com.jrjr.inbest.global.exception.DuplicateException;
 import com.jrjr.inbest.jwt.repository.RefreshTokenRepository;
 import com.jrjr.inbest.login.dto.LoginDto;
 import com.jrjr.inbest.login.entity.Login;
-import com.jrjr.inbest.login.exception.AuthenticationFailedException;
-import com.jrjr.inbest.login.exception.DuplicateException;
 import com.jrjr.inbest.login.repository.LoginRepository;
 import com.jrjr.inbest.user.entity.User;
 import com.jrjr.inbest.user.repository.UserRepository;
@@ -73,6 +73,8 @@ public class LoginServiceImpl implements LoginService {
 		}
 
 		// redis 에서 refreshToken 삭제
-		refreshTokenRepository.deleteById(inputLoginDto.getEmail());
+		if (refreshTokenRepository.existsById(inputLoginDto.getEmail())) {
+			refreshTokenRepository.deleteById(inputLoginDto.getEmail());
+		}
 	}
 }
